move everything to typescript directory Signed-off-by: piotr.swierzy <piotr.swierzy@arianelabs.com>
