+++ typescript/src/langchain/tool.ts
