+++ typescript/src/langchain/toolkit.ts
@@ -2,7 +2,7 @@ import { BaseToolkit } from '@langchain/core/tools';
 import HederaAgentKitTool from './tool';
 import HederaAgentKitAPI from '../shared/api';
 import tools from '../shared/tools';
-import { type Configuration } from '@/shared/configuration';
+import { type Configuration } from 'typescript/src/shared/configuration';
 import { Client } from '@hashgraph/sdk';
 
 class HederaLangchainToolkit implements BaseToolkit {
