+++ typescript/src/shared/configuration.ts
