+++ typescript/src/shared/tools.ts
