+++ typescript/src/shared/tools/consensus/create-topic.ts
@@ -4,10 +4,10 @@ import type { Tool } from '../../tools';
 import { Client } from '@hashgraph/sdk';
 import { handleTransaction } from '../../strategies/tx-mode-strategy';
 import HederaBuilder from '../../hedera-utils/hedera-builder';
-import { createTopicParameters } from '@/shared/parameter-schemas/hcs.zod';
-import HederaParameterNormaliser from '@/shared/hedera-utils/hedera-parameter-normaliser';
-import { getMirrornodeService } from '@/shared/hedera-utils/mirrornode/hedera-mirrornode-utils';
-import { IHederaMirrornodeService } from '@/shared/hedera-utils/mirrornode/hedera-mirrornode-service.interface';
+import { createTopicParameters } from 'typescript/src/shared/parameter-schemas/hcs.zod';
+import HederaParameterNormaliser from 'typescript/src/shared/hedera-utils/hedera-parameter-normaliser';
+import { getMirrornodeService } from 'typescript/src/shared/hedera-utils/mirrornode/hedera-mirrornode-utils';
+import { IHederaMirrornodeService } from 'typescript/src/shared/hedera-utils/mirrornode/hedera-mirrornode-service.interface';
 
 const createTopicPrompt = (_context: Context = {}) => `
 This tool will create a new topic on the Hedera network.
