+++ typescript/src/shared/tools/consensus/submit-topic-message.ts
@@ -4,8 +4,8 @@ import type { Tool } from '../../tools';
 import { Client } from '@hashgraph/sdk';
 import { handleTransaction } from '../../strategies/tx-mode-strategy';
 import HederaBuilder from '../../hedera-utils/hedera-builder';
-import { submitTopicMessageParameters } from '@/shared/parameter-schemas/hcs.zod';
-import HederaParameterNormaliser from '@/shared/hedera-utils/hedera-parameter-normaliser';
+import { submitTopicMessageParameters } from 'typescript/src/shared/parameter-schemas/hcs.zod';
+import HederaParameterNormaliser from 'typescript/src/shared/hedera-utils/hedera-parameter-normaliser';
 
 const submitTopicMessagePrompt = (_context: Context = {}) => `
 This tool will submit a message to a topic on the Hedera network.
