+++ typescript/src/shared/tools/queries/get-account-query.ts
