+++ typescript/src/shared/tools/types.ts
