+++ typescript/tsconfig.json
@@ -4,7 +4,11 @@
     "compilerOptions": {
         "target": "ES2020",
         "module": "ESNext",
-        "lib": ["ES2020", "DOM", "DOM.Iterable"],
+        "lib": [
+            "ES2020",
+            "DOM",
+            "DOM.Iterable"
+        ],
         "strict": true,
         "noImplicitAny": true,
         "strictBindCallApply": true,
@@ -27,7 +31,7 @@
         "baseUrl": ".",
         "paths": {
             "@/*": [
-                "src/*"
+                "typescript/src/*"
             ]
         }
     },
@@ -35,8 +39,8 @@
         "**/*.ts"
     ],
     "exclude": [
-        "node_modules",
+        "typescript/node_modules",
         "examples",
-        "dist"
+        "typescript/dist"
     ]
 }
